fix(i18n): fix next-i18next in static pages #2
fix #2
