+++ next-i18next.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   i18n: {
     defaultLocale: 'zh',
-    locales: ['en', 'zh'],
+    locales: ['en', 'zh']
   },
 }
