+++ next.config.js
@@ -1,13 +1,11 @@
 /** @type {import('next').NextConfig} */
-const { i18n } = require('./next-i18next.config')
 
 const nextConfig = {
   reactStrictMode: true,
   images: {
     loader: 'akamai',
-    path: '',
+    path: '/',
   },
-  i18n,
 }
 
 module.exports = nextConfig
