+++ src/pages/index.tsx
@@ -1,37 +1,35 @@
-import MainLayout from '@/layouts/MainLayout';
-import Head from 'next/head';
-import { ReactElement } from 'react';
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import { useTranslation } from 'next-i18next';
-import Developing from '@/components/Developing';
+import i18nextConfig from "@/../next-i18next.config";
+import { useRouter } from "next/router"
+import { useEffect } from "react";
 
-export default function Home() {
-
-  const { t } = useTranslation('common');
-
-  return (
-    <>
-      <Head>
-        <title>{t('title', { title: t('home') })}</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Developing />
-    </>
-  )
+interface IndexProps {
+  locales: string[]
 }
 
-Home.getLayout = (page: ReactElement) => {
-  return (
-    <MainLayout>{page}</MainLayout>
-  )
-};
+export default function Index({ locales }: IndexProps) {
+  const router = useRouter();
 
-export async function getStaticProps(context: { locale: string }) {
+  useEffect(() => {
+    for (const lang of navigator.languages) {
+      for (const locale of locales) {
+        if (lang.startsWith(locale)) {
+          const path = '/' + locale;
+          router.replace(path);
+          return;
+        }
+      }
+    }
+  }, []);
+
+  return <></>;
+}
+
+export function getStaticProps() {
+  const { locales } = i18nextConfig.i18n;
+
   return {
     props: {
-      ...(await serverSideTranslations(context.locale, ['common'])),
+      locales
     },
   }
 }
